+++ resources/views/article/show.blade.php
@@ -60,8 +60,10 @@
       <div class="card-body">
         <h5 class="card-title">{{$comment->name}}</h5>
         <p class="card-text">{{$comment->desc}}</p>
-        <a href="/comment/{{$comment->id}}/edit" class="btn btn-primary">Edit comment</a>
-        <a href="/comment/{{$comment->id}}/delete" class="btn btn-warning">Delete comment</a>
+        @can('update_comment',$comment)
+          <a href="/comment/{{$comment->id}}/edit" class="btn btn-primary">Edit comment</a>
+          <a href="/comment/{{$comment->id}}/delete" class="btn btn-warning">Delete comment</a>
+        @endcan
       </div>
     </div>
   </div>
